feat(lib/getIconSvg) move to actions
