--- lib/getIconSvg/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./getIconSvg";
-export { type Svgs } from "./getIconSvg";
